% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fontscales-classes.R, R/stateface.R
\docType{data}
\name{fontscales-extensions}
\alias{fontscales-extensions}
\alias{GeomStateFace}
\title{fontscales extensions to ggplot2}
\description{
fontscales makes heavy use of the fontscales class system to extend the
functionality of ggplot2. In general the actual classes should be of little
interest to users as the standard ggplot2 api of using geom_* and stat_*
functions for building up the plot is encouraged.
}
\keyword{datasets}
